Backend/requirements.txt
flask==2.3.3
torch>=2.1.0
torch_geometric==2.4.0
networkx==3.1
pandas
numpy
python-dotenv==1.0.0
groq==0.4.0
scikit-learn==1.3.0
